+++ app/controllers/v1/users_update_controller.rb
@@ -1,0 +1,62 @@
+# frozen_string_literal: true
+
+# Conovel API
+#
+# No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
+#
+# The version of the OpenAPI document: 1.0.0
+# Generated by: https://github.com/openapitools/openapi-generator.git
+#
+
+module V1
+  # UsersController
+  class UsersUpdateController < ApplicationController
+    include ErrorResponseHelper
+
+    # rubocop:disable Metrics/AbcSize
+    def update_user_by_me
+      # パラメータの存在チェック
+      required_keys = %w[penName nickName isAnonymous birthYm agreedTermsVersion]
+      return unless check_required_keys(params, required_keys)
+
+      user = User.find_by!(user_id: @current_user_id)
+      Rails.logger.debug("[DEBUG]カレントユーザー情報 - user: #{user.to_json}")
+
+      # パラメータをキャメルケースからスネークケースに変換
+      transformed_params = params.transform_keys(&:underscore)
+      Rails.logger.debug("[DEBUG] 変換前のパラメータ: #{params.to_json}")
+      Rails.logger.debug("[DEBUG] 変換後のパラメータ: #{transformed_params.to_json}")
+
+      if user.update!(transformed_params.permit(:pen_name, :nick_name, :is_anonymous, :profile_icon_image, :birth_ym,
+                                                :agreed_terms_version, :remarks))
+        user.reload # 最新状態取得
+        evaluation_good_count = Evaluation.joins(:sentence)
+                                          .where(sentences: { sentence_user_id: user.user_id })
+                                          .where(evaluation: 'good')
+                                          .count
+        Rails.logger.debug("[DEBUG] Evaluation count query result: #{evaluation_good_count}")
+
+        render json: {
+          userId: user.user_id,
+          penName: user.pen_name,
+          nickName: user.nick_name,
+          profileIconImage: user.profile_icon_image,
+          evaluationGoodCount: evaluation_good_count,
+          createdAt: user.created_at,
+          updatedAt: user.updated_at,
+          birthYm: user.birth_ym,
+          isAnonymous: user.is_anonymous,
+          agreedTermsVersion: user.agreed_terms_version
+        }, status: :ok
+      end
+    end
+    # rubocop:enable Metrics/AbcSize
+
+    private
+
+    # カスタムエラーメッセージを定義
+    def custom_record_invalid_message(exception)
+      "ユーザーアカウント情報の更新に失敗しました。: #{exception.record.errors.full_messages.join(', ')}"
+    end
+  end
+end
